+++ lib/lightning_web/endpoint.ex
@@ -44,7 +44,11 @@ defmodule LightningWeb.Endpoint do
     cookie_key: "request_logger"
 
   plug Plug.RequestId
-  plug PromEx.Plug, prom_ex_module: Lightning.PromEx
+
+  plug Unplug,
+    if: {LightningWeb.PromExPlugAuthorization, nil},
+    do: {PromEx.Plug, prom_ex_module: Lightning.PromEx}
+
   plug Plug.Telemetry, event_prefix: [:phoenix, :endpoint]
 
   plug Plugs.WebhookAuth